BE [REFACTOR] - Change time handling
Now the system accepts directly the timestamp. In the request body you must be include the expiration time of the link. This timestamp will be saved in the database with the timezone in order to avoid possibile error

BREAKING CHANGE: Replaced the 'TimeToLive' field of the 'shorten' request body with the 'ExpirationTime' one.