BE [REFACTOR] - Change database connection managing
In order to a great clarity of the code, now the struct that contains the database instance has another field representing the table. In this way it is not necessary to pass it each time you call a database function. Furthermore, the database initialization change to a simple 'New<type>DTO' constructor function